backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.2.2
huggingface_hub==0.36.2
